+++ CHANGELOG.md
@@ -5,17 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.7.0] - 2022/09/17
+## [0.7.0] - 2022/09/23
 ### Added
 - Add `minify` argument to all `getBuffer()` methods.
-- Add `writeBinary` property to CombinedTuningResource.
 - Add `combine()` and `combineAsync()` methods to CombinedTuningResource class.
 - Add `combineTuning()` and `merge()` methods to Package class.
 - Add `merge()` method to StringTableResource class.
 ### Changed
 - Update dependencies on @s4tk/xml-dom and @s4tk/images.
 - CombinedTuningResource no longer throws on clone, equality check, or serialization.
-- CombinedTuningResource.isXml() is now true when `writeBinary === false`.
+- CombinedTuningResource.isXml() is now true, since XML is the only supported format.
 - Minor optimization when loading resources in packages.
 ### Fixed
 - Unhandled exceptions will no longer output squish.js to the console.